Create Service for Model
เพิ่ม Service สำหรับทุก Model

+++ src/main/java/com/fitness/cs/fitness_project_be/repository/TrainerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.fitness.cs.fitness_project_be.model.Trainer;
 
 @Repository
-public interface TrainerRepository extends JpaRepository<Trainer, Long> {
+public interface TrainerRepository extends JpaRepository<Trainer, Integer> {
     
 }
